+++ libs/spark/src/Unstable_List/Unstable_List.tsx
@@ -4,7 +4,7 @@ import Unstable_ListSubheader, {
   Unstable_ListSubheaderProps,
 } from '../Unstable_ListSubheader';
 import { OverridableComponent, OverrideProps, useId } from '../utils';
-import makeStyles from '../makeStyles';
+import withStyles from '../withStyles';
 
 export interface Unstable_ListTypeMap<
   // eslint-disable-next-line @typescript-eslint/ban-types
@@ -29,18 +29,17 @@ export interface Unstable_ListTypeMap<
   classKey: Unstable_ListClassKey;
 }
 
-export type Unstable_ListClassKey =
-  | 'root'
-  | 'rootWithSubheader'
-  | 'rootWithPadding';
-
 export type Unstable_ListProps<
   D extends ElementType = Unstable_ListTypeMap['defaultComponent'],
   // eslint-disable-next-line @typescript-eslint/ban-types
   P = {}
 > = OverrideProps<Unstable_ListTypeMap<P, D>, D>;
 
-const useStyles = makeStyles<Unstable_ListClassKey>(
+export type Unstable_ListClassKey = 'root';
+
+type PrivateClassKey = 'private-root-subheader' | 'private-root-disablePadding';
+
+const withClasses = withStyles<Unstable_ListClassKey | PrivateClassKey>(
   {
     /* Styles applied to the root element. */
     root: {
@@ -48,15 +47,16 @@ const useStyles = makeStyles<Unstable_ListClassKey>(
       margin: 0,
       padding: 0,
       position: 'relative',
-    },
-    /* Styles applied to the root element when a `subheader` is provided. */
-    rootWithSubheader: {
-      paddingBlockStart: 0,
-    },
-    /* Styles applied to the root element if `disablePadding={false}`. */
-    rootWithPadding: {
       paddingBlockStart: 8,
       paddingBlockEnd: 8,
+    },
+    /* Private */
+    'private-root-subheader': {
+      paddingBlockStart: 0,
+    },
+    'private-root-disablePadding': {
+      paddingBlockStart: 0,
+      paddingBlockEnd: 0,
     },
   },
   { name: 'MuiSparkUnstable_List' }
@@ -65,7 +65,7 @@ const Unstable_List: OverridableComponent<Unstable_ListTypeMap> = forwardRef(
   function Unstable_List(props, ref) {
     const {
       children,
-      classes: classesProp,
+      classes,
       className,
       // @ts-expect-error not picked up
       component: Component = 'ul',
@@ -75,24 +75,15 @@ const Unstable_List: OverridableComponent<Unstable_ListTypeMap> = forwardRef(
       ...other
     } = props;
 
-    const classes = useStyles();
-
     const subheaderId = useId(ListSubheaderProps?.id);
 
     return (
       <Component
         className={clsx(
           classes.root,
-          classesProp?.root,
           {
-            [clsx(
-              classes.rootWithSubheader,
-              classesProp?.rootWithSubheader
-            )]: !!subheader,
-            [clsx(
-              classes.rootWithPadding,
-              classesProp?.rootWithPadding
-            )]: !disablePadding,
+            [classes['private-root-subheader']]: !!subheader,
+            [classes['private-root-disablePadding']]: disablePadding,
           },
           className
         )}
@@ -111,4 +102,4 @@ const Unstable_List: OverridableComponent<Unstable_ListTypeMap> = forwardRef(
   }
 );
 
-export default Unstable_List;
+export default withClasses(Unstable_List) as typeof Unstable_List;